Add files via upload
Script to collect images of die or reticle field that can then be used to stitch together a single image of the area using Fiji Stitching function or other panoramic image stitching program. 

The optics in the Summit have uncorrected distortion so it is generally wise to over sample the area to compensate for aberration and distortion.